rename-to-agent: rename to agent_book, xfiles_sorter, agent_notes, jamesbot
