--- address_book/address_book/__init__.py
@@ -1,5 +1,0 @@
-from address_book.entities import AddressBook, Record, DATE_FORMAT
-from address_book.iterators import AddressBookIterator, PaginatedAddressBookIterator
-from address_book.exceptions import CustomExceptions
-
-__all__ = ['AddressBook', 'Record', 'DATE_FORMAT', 'AddressBookIterator', 'PaginatedAddressBookIterator', 'CustomExceptions']
