+++ agent_notes/setup.py
